requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
tensorflow==2.16.1  # Upgrade from 2.10.0
numpy==1.26.4       # Update for TensorFlow 2.16.1 compatibility
pillow==10.4.0
python-multipart==0.0.9